requirements.txt
appnope==0.1.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.0
bleach==1.5.0
bs4==0.0.1
certifi==2016.2.28
cffi==1.10.0
ciseau==1.0.0
configparser==3.5.0
cycler==0.10.0
cymem==1.31.2
Cython==0.22
cytoolz==0.8.2
decorator==4.1.2
dill==0.2.7.1
entrypoints==0.2.3
enum34==1.1.6
ftfy==4.4.3
funcsigs==1.0.2
functools32==3.2.3.post2
html5lib==0.9999999
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
Jinja2==2.9.6
joblib==0.8.4
jsonschema==2.6.0
lxml==3.8.0
MarkupSafe==1.0
mistune==0.7.4
mock==2.0.0
murmurhash==0.26.4
networkx==1.11
nose==1.3.7
numpy==1.13.1
olefile==0.44
pandocfilters==1.4.2
pathlib==1.0.1
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==2.8.1
plac==0.9.6
preshed==1.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.4.34
pycparser==2.18
Pygments==2.2.0
pyparsing==2.2.0
PySide==1.2.2
pytest==2.7.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.11
pyzmq==16.0.2
qtconsole==4.3.1
regex==2017.7.28
requests==2.14.2
scandir==1.5
scikit-learn==0.16.0
scipy==0.15.1
segment==0.2.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
subprocess32==3.2.7
termcolor==1.1.0
terminado==0.6
testpath==0.3
thinc==6.5.2
toolz==0.8.2
torchvision==0.1.9
tornado==4.5.2
tqdm==4.15.0
traitlets==4.3.2
ujson==1.35
wcwidth==0.1.7
widgetsnbextension==3.0.2
wrapt==1.10.11
